finished MMU testing

+++ include/utility/heap.h
@@ -74,15 +74,15 @@ public:
     }
 
     static void typed_free(void * ptr) {
-        int * addr = reinterpret_cast<int *>(ptr);
-        unsigned int bytes = *--addr;
+        long * addr = reinterpret_cast<long *>(ptr);
+        unsigned long bytes = *--addr;
         Heap * heap = reinterpret_cast<Heap *>(*--addr);
         heap->free(addr, bytes);
     }
 
     static void untyped_free(Heap * heap, void * ptr) {
-        int * addr = reinterpret_cast<int *>(ptr);
-        unsigned int bytes = *--addr;
+        long * addr = reinterpret_cast<long *>(ptr);
+        unsigned long bytes = *--addr;
         heap->free(addr, bytes);
     }
